requirements.txt
cycler==0.10.0
decorator==4.4.2
filterpy==1.4.5
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
matplotlib==3.3.3
networkx==2.5
numba==0.52.0
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.22.2
scipy==1.6.0
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.1.8
torch==1.7.1
torchvision==0.8.2
tqdm==4.55.2
typing-extensions==3.7.4.3